+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add a node ip and node start time field to the node info struct.
+
 ## [0.7.7] - 2026-05-06
 
 ### Added